+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.2.1
 
 - Have examples/timer.rb show `msg_box` on finish
+- Have examples/color_the_circles.rb push colored circles when colored behind uncolored circles to keep uncolored circles visible
 
 ## 0.2.0